fix(rustffi): Guard Rust FFI behind build tag to fix CI linker errors
Move rustffi package behind `rust_ffi` build tag so CI builds without
the Rust library don't fail with `-ldefra_ffi` linker errors. Also fix
JS/wasm build by syncing db_setup_js.go setupNode signature, and remove
cbindings dependency on Rust extern by moving RegisterIdentityWithRust
into the rustffi package where it belongs.
