+++ cbindings/identity.go
@@ -13,55 +13,16 @@ package cbindings
 /*
 #include <stdlib.h>
 #include "defra_structs.h"
-extern Result RegisterIdentity(char* did, char* privateKeyHex, char* publicKeyHex, char* keyType);
 */
 import "C"
 
 import (
 	"context"
-	"encoding/hex"
 	"runtime/cgo"
-	"unsafe"
 
 	"github.com/sourcenetwork/defradb/acp/identity"
 	"github.com/sourcenetwork/defradb/crypto"
 )
-
-// RegisterIdentityWithRust registers an identity's signing config with Rust's global identity store.
-// This allows Go-created identities to be used for block signing in Rust.
-func RegisterIdentityWithRust(ident identity.Identity) error {
-	if ident == nil {
-		return nil
-	}
-
-	// Only full identities have private keys and can be used for signing
-	fullIdent, ok := ident.(identity.FullIdentity)
-	if !ok {
-		// Not a full identity, can't register for signing (no private key)
-		return nil
-	}
-
-	did := ident.DID()
-	privateKeyHex := hex.EncodeToString(fullIdent.PrivateKey().Raw())
-	publicKeyHex := hex.EncodeToString(ident.PublicKey().Raw())
-	keyType := string(fullIdent.PrivateKey().Type())
-
-	cDid := C.CString(did)
-	cPrivateKeyHex := C.CString(privateKeyHex)
-	cPublicKeyHex := C.CString(publicKeyHex)
-	cKeyType := C.CString(keyType)
-	defer C.free(unsafe.Pointer(cDid))
-	defer C.free(unsafe.Pointer(cPrivateKeyHex))
-	defer C.free(unsafe.Pointer(cPublicKeyHex))
-	defer C.free(unsafe.Pointer(cKeyType))
-
-	result := C.RegisterIdentity(cDid, cPrivateKeyHex, cPublicKeyHex, cKeyType)
-	res := ConvertAndFreeCResult(result)
-	if res.Status != 0 {
-		return NewErrFFI(res.Error)
-	}
-	return nil
-}
 
 //export NewIdentity
 func NewIdentity(keyType *C.char) C.NewIdentityResult {
@@ -75,12 +36,6 @@ func NewIdentity(keyType *C.char) C.NewIdentityResult {
 	if err != nil {
 		return returnNewIdentityResultC(1, err.Error(), nil)
 	}
-
-	// Register the identity with Rust so block signing works
-	if err := RegisterIdentityWithRust(newIdentity); err != nil {
-		return returnNewIdentityResultC(1, "failed to register identity with Rust: "+err.Error(), nil)
-	}
-
 	return returnNewIdentityResultC(0, "", newIdentity)
 }
