+++ tests/integration/client_setup_rustffi.go
@@ -1,0 +1,132 @@
+// Copyright 2025 Democratized Data Foundation
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+//go:build !js && rust_ffi
+
+package tests
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/sourcenetwork/immutable"
+
+	acpIdentity "github.com/sourcenetwork/defradb/acp/identity"
+	"github.com/sourcenetwork/defradb/node"
+	"github.com/sourcenetwork/defradb/tests/clients"
+	"github.com/sourcenetwork/defradb/tests/clients/rustffi"
+	"github.com/sourcenetwork/defradb/tests/state"
+)
+
+// setupRustFFIClient creates a Rust FFI wrapper and mirrors the Go node's NAC
+// state onto it. This ensures the Rust FFI node has the same access control
+// configuration as the Go node, matching Go's initializeNodeACP() flow.
+func setupRustFFIClient(
+	s *state.State,
+	nodeObj *node.Node,
+	identity immutable.Option[acpIdentity.Identity],
+	nodeIndex int,
+	enableSigning bool,
+	dbPath string,
+) (clients.Client, error) {
+	var wrapper *rustffi.Wrapper
+	var err error
+
+	// Get the node identity for this node index - this is the same identity
+	// used for signing in the Go node (set via db.WithNodeIdentity)
+	var nodeIdentity acpIdentity.Identity
+	if enableSigning {
+		nodeIdentity = state.GetIdentity(s, NodeIdentity(nodeIndex))
+	}
+
+	// For file-based storage, use a subdirectory so Rust's redb doesn't collide
+	// with Go's badger store at the same path.
+	rustDBPath := ""
+	if dbPath != "" {
+		rustDBPath = dbPath + "/rustffi"
+	}
+
+	// Build SourceHub config if SourceHub ACP is active.
+	// GRPCAddress here is actually the REST/LCD API address (the Rust client uses REST, not gRPC).
+	var shConfig *rustffi.SourceHubConfig
+	if s.DocumentACPType == state.SourceHubDocumentACPType && s.SourcehubRestAddress != "" {
+		shConfig = &rustffi.SourceHubConfig{
+			GRPCAddress:     s.SourcehubRestAddress,
+			CometRPCAddress: s.SourcehubCometRPCAddress,
+			ChainID:         s.SourcehubChainID,
+			SignerKey:       s.SourcehubSignerKey,
+		}
+	}
+
+	if s.IsNetworkEnabled {
+		listenAddr := "/ip4/" + getIPString() + "/tcp/0"
+		wrapper, err = rustffi.NewWrapperWithP2P(listenAddr, enableSigning, nodeIdentity, rustDBPath, shConfig)
+	} else {
+		wrapper, err = rustffi.NewWrapper(enableSigning, nodeIdentity, rustDBPath, shConfig)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	// Forward SE artifact events from Go's event bus to the wrapper's event bus.
+	// Go's SE coordinator publishes to Go's event bus, but the test framework
+	// subscribes to the wrapper's event bus via c.Events().
+	if err := wrapper.ForwardSEEvents(nodeObj.DB.Events()); err != nil {
+		wrapper.Close()
+		return nil, fmt.Errorf("failed to forward SE events: %w", err)
+	}
+
+	// Mirror the Go node's NAC state onto the Rust FFI node.
+	nacStatus, nacErr := nodeObj.DB.GetNACStatus(s.Ctx)
+	if nacErr == nil && identity.HasValue() {
+		ownerDID := identity.Value().DID()
+		if ownerDID != "" && (nacStatus.Status == "enabled" || nacStatus.Status == "disabled temporarily") {
+			// Check if Rust node already has NAC enabled (from persistent store)
+			rustStatus, rustErr := wrapper.GetNACStatus(s.Ctx)
+			needsEnable := rustErr != nil || rustStatus.Status == "not configured"
+
+			if needsEnable {
+				if err := wrapper.EnableNACForInit(ownerDID); err != nil {
+					wrapper.Close()
+					return nil, fmt.Errorf("failed to enable NAC on Rust FFI node: %w", err)
+				}
+			}
+			// Mirror Go's nodeIdentity shortcut
+			nodeIdentityDID := getIdentityDID(s, NodeIdentity(nodeIndex))
+			rustStatusForAdmin, _ := wrapper.GetNACStatus(s.Ctx)
+			if nodeIdentityDID != "" && nodeIdentityDID != ownerDID && rustStatusForAdmin.Status != "disabled temporarily" {
+				if err := wrapper.AddNACAdminForInit(ownerDID, nodeIdentityDID); err != nil {
+					wrapper.Close()
+					return nil, fmt.Errorf("failed to add node identity as NAC admin: %w", err)
+				}
+			}
+
+			// If Go says "disabled temporarily", mirror that state on Rust
+			if nacStatus.Status == "disabled temporarily" {
+				rustStatus2, rustErr2 := wrapper.GetNACStatus(s.Ctx)
+				if rustErr2 != nil || rustStatus2.Status != "disabled temporarily" {
+					ctx := acpIdentity.WithContext(s.Ctx, identity)
+					if err := wrapper.DisableNAC(ctx); err != nil {
+						wrapper.Close()
+						return nil, fmt.Errorf("failed to mirror NAC disabled state: %w", err)
+					}
+				}
+			}
+		}
+	}
+
+	// Store the Go node closer so the Go node's badger lock is released
+	// when the wrapper is closed during restart tests.
+	wrapper.SetGoNodeCloser(func() {
+		_ = nodeObj.Close(context.Background())
+	})
+
+	return wrapper, nil
+}
